requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
python-socketio==5.4.0
python-engineio==4.2.1
aiortc==1.2.1
numpy
gevent==21.8.0
gevent-websocket==0.10.1